another go at default typing, list of lists is treated as its own case

+++ tests/test_image_io.py
@@ -41,7 +41,7 @@ class TestImage(unittest.TestCase):
         img = Image(im)
         self.assertIsInstance(img, Image)
         self.assertEqual(img.shape, im.shape)
-        self.assertEqual(img.dtype, np.float32)
+        self.assertEqual(img.dtype, np.float64)
 
         # set type as float, then make sure isimage is true
         img = Image(im.astype(np.float32))
